+++ template/src/index.tsx
@@ -1,7 +1,7 @@
 import React from 'react'
 import {render} from 'react-dom'
 import {Provider} from '@tesler-ui/core'
-import {LocaleProvider} from 'antd'
+import {ConfigProvider} from 'antd'
 import enUs from 'antd/es/locale-provider/en_US'
 import {reducers} from './reducers'
 import {epics} from './epics'
@@ -14,9 +14,9 @@ const App = <Provider
     customEpics={epics}
     axiosInstance={axiosInstance}
 >
-    <LocaleProvider locale={enUs}>
+    <ConfigProvider locale={enUs}>
         <AppLayout />
-    </LocaleProvider>
+    </ConfigProvider>
 </Provider>
 
 render(App, document.getElementById('root'))